chore: create separate class for GroupEvent class and add documentation
